+++ CHANGES.rst
@@ -1208,6 +1208,10 @@ tweakreg
 - Added the 'GAIADR3' catalog to the available options for alignment;
   this has been enabled as the default option [#7611].
 
+- Apply median filtering to IVM weight array to better handle saturated
+  pixels. Added option of 'ivm-med', 'ivm-smed' (recommended default) to the
+  ``weight_type`` argument of ``ResampleStep``. [#7563]
+
 
 1.10.2 (2023-04-14)
 ===================